requirements.txt
subprocess   
sqlite3
webbrowser
pywebview